No longer sleep after sending fake backspaces.
We used to wait a few ms after sending fake backspaces to ensure they
are properly processed. That's because on GTK, fake backspaces
are sent as XIM event that will later be sent back into IBus where
they would be skipped and finally sent to the editing GTK app.
Whereas ibus_commit's message will be sent directly to the app.
Therefore the events aren't in sync.

Now we force GTK to use XIM as the IMModule so all the events
are synchronous.