Handle "uơ" in "huơ", "thuở", "quở"
If the current word has no last consonant and the first consonant
is one of "h", "th" and the vowel is "ươ" then change the vowel into
"uơ", keeping case and accent. If an alphabet character is then added
into the word then change back to "ươ".

NOTE: In the dictionary, these are the only words having this strange
vowel so we don't need to worry about other cases.