Make ENTER behave normally if no choice was made
If the user hasn't made any choice (they agree with the default
one) then it doesn't make sense for them to have to press ENTER
to choose what's already chosen. It's annoying having to press
ENTER twice when typing an English word like 'system' *and*
making a new line after it. 

Thus, this commit makes ENTER behave like a normal ENTER if the
default option is untouched. Still, if they have moved the cursor
then they still have to press ENTER to select whether or not the
choice is the default one.